Map remaining core .rodata
