+++ include/sys/crash.h
@@ -8,13 +8,6 @@
 #include "PR/sched.h"
 #include "sys/dll.h"
 
-// used for crash screens
-typedef struct ErrString {
-	u32 code1;
-	u32 code2;
-	char *text;
-} ErrString;
-
 // Length of gCrashMesgQueueBuffer
 #define CRASH_MESG_QUEUE_BUFFER_LENGTH 1
