+++ splat.yaml
@@ -41,7 +41,7 @@ segments:
     - [0x1C40, c, boot]
     - [0x1E20, c, camera]
     - [0x5660, c, curves]
-    - [0x7B50, c, segment_7B50]
+    - [0x7B50, c, voxmap]
     - [0xB9F0, c, generic_queue]
     - [0xBC10, c, generic_stack]
     - [0xBED0, c, segment_BED0]
@@ -53,7 +53,7 @@ segments:
     - [0x10DF0, c, input]
     - [0x11EF0, c, segment_11EF0]
     - [0x11F70, c, segment_11F70]
-    - [0x12320, c, segment_12320]
+    - [0x12320, c, audio]
     - [0x13030, c, mpeg_fs]
     - [0x130E0, c, queue]
     - [0x13D70, c, main]
@@ -67,7 +67,7 @@ segments:
     - [0x20490, c, segment_20490]
     - [0x21150, c, object]
     - [0x24930, c, objanim]
-    - [0x26900, c, segment_26900]
+    - [0x26900, c, objhits]
     - [0x31AC0, c, objtype]
     - [0x32220, c, objmsg]
     - [0x326A0, c, segment_326A0]
@@ -82,12 +82,14 @@ segments:
     - [0x3C1B0, c, scheduler]
     - [0x3D850, c, segment_3D850]
     - [0x3D890, c, texture]
+    - [0x3F5F0, c, segment_3F5F0]
     - [0x41BD0, c, map]
     - [0x4E070, c, newshadows]
     - [0x535C0, c, segment_535C0]
     - [0x53F00, c, segment_53F00] # track/intersect.c ?
     - [0x5C470, c, segment_5C470]
     - [0x5D380, c, video]
+    - [0x5E010, c, vi]
     - [0x5EE00, c, libultra/io/vimgr]
     - [0x5F150, c, libultra/gu/perspective]
     - [0x5F3D0, c, libultra/gu/ortho]
@@ -317,7 +319,7 @@ segments:
     - [0x93CB0, .data, libultra/io/vimgr]
     - [0x93CE0, .data, print]
     - [0x93DF0, .data, debug]
-    - [0x93E00, data, exception]
+    - [0x93E00, .data, exception]
     - [0x94410, data, segment_63EB0]
     - [0x94490, .data, di_comm]
     - [0x944B0, .data, libnaudio/n_sl]
@@ -358,36 +360,44 @@ segments:
     - [0x98E60, .rodata, boot]
     - [0x98E90, .rodata, camera]
     - [0x98FB0, .rodata, curves]
-    - [0x99110, rodatabin]
+    - [0x99110, .rodata, voxmap]
     - [0x992D0, .rodata, bitstream]
     - [0x99330, .rodata, dll]
     - [0x99470, .rodata, fonts]
     - [0x99C30, .rodata, menu]
-    - [0x99C60, rodatabin]
+    - [0x99C60, .rodata, audio]
     - [0x99CA0, .rodata, queue]
     - [0x99D10, .rodata, main]
     - [0x99DB0, rodata, math_util]
     - [0x99DE0, .rodata, memory]
-    - [0x99E40, rodatabin]
+    - [0x99E40, .rodata, model]
+    - [0x9A0D0, .rodata, segment_1D900]
+    - [0x9A0E0, .rodata, object]
     - [0x9a400, .rodata, objanim]
-    - [0x9a460, rodatabin]
+    - [0x9a460, .rodata, objhits]
     - [0x9A700, .rodata, objtype]
     - [0x9A850, .rodata, objmsg]
-    - [0x9A890, rodatabin]
+    - [0x9A890, .rodata, segment_334F0]
+    - [0x9A950, .rodata, objprint]
+    - [0x9A9B0, .rodata, segment_3A190]
     - [0x9AAB0, .rodata, rarezip]
     - [0x9AB30, rodata, rarezip_asm]
     - [0x9AC20, .rodata, scheduler]
-    - [0x9AF70, rodatabin]
+    - [0x9AF70, .rodata, texture]
+    - [0x9B080, .rodata, segment_3F5F0]
     - [0x9B110, .rodata, map]
     - [0x9B5D0, .rodata, newshadows]
-    - [0x9B650, rodatabin]
+    - [0x9B650, rodatabin] # TODO: this is for segment_535C0 but .late_rodata migration isn't working
+    - [0x9B670, .rodata, segment_53F00]
+    - [0x9B870, .rodata, segment_5C470]
     - [0x9B8C0, .rodata, video]
-    - [0x9B8E0, rodatabin]
+    - [0x9B8E0, .rodata, vi]
     - [0x9B980, .rodata, libultra/gu/perspective]
     - [0x9B990, .rodata, assert]
     - [0x9B9F0, .rodata, print]
     - [0x9BC70, .rodata, debug]
-    - [0x9BC90, rodatabin, exception_rodata]  # might be a split at 0x8009b670
+    - [0x9BC90, .rodata, exception]
+    - [0x9C270, .rodata, dl_debug]
     - [0x9C2A0, rodata, developer_names]
     - [0x9C320, .rodata, di_comm]
     - [0x9C420, .rodata, libnaudio/n_synthesizer]
