+++ src/exception.c
@@ -1,9 +1,157 @@
 #include "PR/os_internal.h"
-#include "common.h"
 #include "sys/crash.h"
 #include "libc/stdarg.h"
+#include "common.h"
 
-void get_err_string(s32 x, s32 y, u32 param3, ErrString *param4);
+typedef struct {
+    u32 code1;
+    u32 code2;
+    char *text;
+} CrashErrString;
+
+char D_80093200[] = { 
+    0x0c, 0x0c, 0x0c, 0x00, 0x0c, 0x33, 0x33, 0x00, 0x00, 0x00, 0x26, 0x3f, 0x26, 0x3f, 0x26, 0x2c, 
+    0x0e, 0x2e, 0x2c, 0x0e, 0x33, 0x28, 0x1d, 0x0a, 0x33, 0x1d, 0x33, 0x2d, 0x33, 0x3e, 0x0c, 0x0c, 
+    0x00, 0x00, 0x00, 0x0e, 0x03, 0x03, 0x03, 0x0e, 0x2c, 0x30, 0x30, 0x30, 0x2c, 0x00, 0x0c, 0x3f, 
+    0x1d, 0x37, 0x00, 0x0c, 0x3f, 0x0c, 0x00, 0x00, 0x00, 0x00, 0x0d, 0x03, 0x00, 0x00, 0x3f, 0x00, 
+    0x00, 0x00, 0x00, 0x00, 0x04, 0x0c, 0x30, 0x38, 0x19, 0x0b, 0x03, 0x2e, 0x33, 0x33, 0x33, 0x2e, 
+    0x0c, 0x0f, 0x0c, 0x0c, 0x3f, 0x2f, 0x30, 0x2e, 0x03, 0x3f, 0x3f, 0x30, 0x3f, 0x30, 0x3f, 0x33, 
+    0x33, 0x3f, 0x30, 0x30, 0x3f, 0x03, 0x2e, 0x30, 0x2f, 0x3e, 0x03, 0x2f, 0x33, 0x2f, 0x3f, 0x30, 
+    0x30, 0x30, 0x30, 0x2e, 0x33, 0x2e, 0x33, 0x2e, 0x3e, 0x33, 0x3e, 0x30, 0x2f, 0x00, 0x0c, 0x00, 
+    0x0c, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+    0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+    0x1d, 0x33, 0x3f, 0x33, 0x33, 0x0f, 0x33, 0x0f, 0x33, 0x0f, 0x1d, 0x33, 0x03, 0x33, 0x1d, 0x0f, 
+    0x33, 0x33, 0x33, 0x0f, 0x3f, 0x03, 0x1f, 0x03, 0x3f, 0x3f, 0x03, 0x1f, 0x03, 0x03, 0x2e, 0x03, 
+    0x33, 0x33, 0x2e, 0x33, 0x33, 0x3f, 0x33, 0x33, 0x3f, 0x0c, 0x0c, 0x0c, 0x3f, 0x3f, 0x30, 0x33, 
+    0x33, 0x2e, 0x33, 0x33, 0x1f, 0x33, 0x33, 0x03, 0x03, 0x03, 0x03, 0x3f, 0x37, 0x3f, 0x37, 0x33, 
+    0x33, 0x37, 0x3b, 0x3f, 0x3b, 0x37, 0x0c, 0x33, 0x33, 0x33, 0x0c, 0x1f, 0x33, 0x1f, 0x03, 0x03, 
+    0x2e, 0x33, 0x33, 0x37, 0x3e, 0x0f, 0x33, 0x1f, 0x33, 0x33, 0x2e, 0x07, 0x2e, 0x34, 0x2e, 0x3f, 
+    0x0c, 0x0c, 0x0c, 0x0c, 0x33, 0x33, 0x33, 0x33, 0x2e, 0x33, 0x33, 0x37, 0x2a, 0x0c, 0x33, 0x33, 
+    0x37, 0x3f, 0x37, 0x33, 0x33, 0x1d, 0x33, 0x33, 0x33, 0x33, 0x2e, 0x0c, 0x0c, 0x3f, 0x24, 0x19, 
+    0x06, 0x3f, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+    0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+    0x1d, 0x33, 0x3f, 0x33, 0x33, 0x0f, 0x33, 0x0f, 0x33, 0x0f, 0x1d, 0x33, 0x03, 0x33, 0x1d, 0x0f, 
+    0x33, 0x33, 0x33, 0x0f, 0x3f, 0x03, 0x1f, 0x03, 0x3f, 0x3f, 0x03, 0x1f, 0x03, 0x03, 0x2e, 0x03, 
+    0x33, 0x33, 0x2e, 0x33, 0x33, 0x3f, 0x33, 0x33, 0x3f, 0x0c, 0x0c, 0x0c, 0x3f, 0x3f, 0x30, 0x33, 
+    0x33, 0x2e, 0x33, 0x33, 0x1f, 0x33, 0x33, 0x03, 0x03, 0x03, 0x03, 0x3f, 0x37, 0x3f, 0x37, 0x33, 
+    0x33, 0x37, 0x3b, 0x3f, 0x3b, 0x37, 0x0c, 0x33, 0x33, 0x33, 0x0c, 0x1f, 0x33, 0x1f, 0x03, 0x03, 
+    0x2e, 0x33, 0x33, 0x37, 0x3e, 0x0f, 0x33, 0x1f, 0x33, 0x33, 0x2e, 0x07, 0x2e, 0x34, 0x2e, 0x3f, 
+    0x0c, 0x0c, 0x0c, 0x0c, 0x33, 0x33, 0x33, 0x33, 0x2e, 0x33, 0x33, 0x37, 0x2a, 0x0c, 0x33, 0x33, 
+    0x37, 0x3f, 0x37, 0x33, 0x33, 0x1d, 0x33, 0x33, 0x33, 0x33, 0x2e, 0x0c, 0x0c, 0x3f, 0x24, 0x19, 
+    0x06, 0x3f
+};
+u16 D_800933C4[] = { 
+    0x0000, 0x5294, 0xa528, 0xfffe, 
+    0x0000, 0x0014, 0x0028, 0x003e, 
+    0x0000, 0x0294, 0x0528, 0x07fe, 
+    0x0000, 0x0280, 0x0500, 0x07c0, 
+    0x0000, 0x5280, 0xa500, 0xffc0, 
+    0x0000, 0x5000, 0xa000, 0xf800, 
+    0x0000, 0x5014, 0xa028, 0xf83e, 
+    0x0000, 0x0000, 0x0000, 0x0000
+};
+CrashErrString errStringArray_cause[] = {
+    { 0x80000000, 0x80000000, "BD" },
+    { 0x00008000, 0x00008000, "IP8" },
+    { 0x00004000, 0x00004000, "IP7" },
+    { 0x00002000, 0x00002000, "IP6" },
+    { 0x00001000, 0x00001000, "IP5" },
+    { 0x00000800, 0x00000800, "IP4" },
+    { 0x00000400, 0x00000400, "IP3" },
+    { 0x00000200, 0x00000200, "IP2" },
+    { 0x00000100, 0x00000100, "IP1" },
+    { 0x0000007C, 0x00000000, "Interrupt" },
+    { 0x0000007C, 0x00000004, "TLB modify ex" },
+    { 0x0000007C, 0x00000008, "TLB load/fetch ex" },
+    { 0x0000007C, 0x0000000C, "TLB store ex" },
+    { 0x0000007C, 0x00000010, "Addr load/fetch err" },
+    { 0x0000007C, 0x00000014, "Addr store err" },
+    { 0x0000007C, 0x00000018, "Bus error fetch ex" },
+    { 0x0000007C, 0x0000001C, "Bus error ref ex" },
+    { 0x0000007C, 0x00000020, "Syscall ex" },
+    { 0x0000007C, 0x00000024, "Breakpoint ex" },
+    { 0x0000007C, 0x00000028, "Reserved inst ex" },
+    { 0x0000007C, 0x0000002C, "Co-pro unusable ex" },
+    { 0x0000007C, 0x00000030, "Arith overflow ex" },
+    { 0x0000007C, 0x00000034, "Trap ex" },
+    { 0x0000007C, 0x00000038, "Virtual coherency ex" },
+    { 0x0000007C, 0x0000003C, "Floating point ex" },
+    { 0x0000007C, 0x0000005C, "Watchpoint ex" },
+    { 0x0000007C, 0x0000007C, "Virtual coherency ex2" },
+    { 0x00000000, 0x00000000, "" },
+};
+CrashErrString errStringArray_cause2[] = {
+    { 0x80000000, 0x80000000, "CU3" },
+    { 0x40000000, 0x40000000, "CU2" },
+    { 0x20000000, 0x20000000, "CU1" },
+    { 0x10000000, 0x10000000, "CU0" },
+    { 0x08000000, 0x08000000, "RP" },
+    { 0x04000000, 0x04000000, "FR" },
+    { 0x02000000, 0x02000000, "RE" },
+    { 0x00400000, 0x00400000, "BEV" },
+    { 0x00200000, 0x00200000, "TS" },
+    { 0x00100000, 0x00100000, "SR" },
+    { 0x00040000, 0x00040000, "CH" },
+    { 0x00020000, 0x00020000, "CE" },
+    { 0x00010000, 0x00010000, "DE" },
+    { 0x00008000, 0x00008000, "IM8" },
+    { 0x00004000, 0x00004000, "IM7" },
+    { 0x00002000, 0x00002000, "IM6" },
+    { 0x00001000, 0x00001000, "IM5" },
+    { 0x00000800, 0x00000800, "IM4" },
+    { 0x00000400, 0x00000400, "IM3" },
+    { 0x00000200, 0x00000200, "IM2" },
+    { 0x00000100, 0x00000100, "IM1" },
+    { 0x00000080, 0x00000080, "KX" },
+    { 0x00000040, 0x00000040, "SX" },
+    { 0x00000020, 0x00000020, "UX" },
+    { 0x00000018, 0x00000010, "USR" },
+    { 0x00000018, 0x00000008, "SUP" },
+    { 0x00000018, 0x00000000, "KER" },
+    { 0x00000004, 0x00000004, "ERL" },
+    { 0x00000002, 0x00000002, "EXL" },
+    { 0x00000001, 0x00000001, "IE" },
+    { 0x00000000, 0x00000000, "" },
+};
+CrashErrString errStringArray_fpsr[] = {
+    { 0x01000000, 0x01000000, "FS" },
+    { 0x00800000, 0x00800000, "C" },
+    { 0x00020000, 0x00020000, "Unimplemented operation" },
+    { 0x00010000, 0x00010000, "Invalid operation" },
+    { 0x00008000, 0x00008000, "Division by zero" },
+    { 0x00004000, 0x00004000, "Overflow" },
+    { 0x00002000, 0x00002000, "Underflow" },
+    { 0x00001000, 0x00001000, "Inexact operation" },
+    { 0x00000800, 0x00000800, "EV" },
+    { 0x00000400, 0x00000400, "EZ" },
+    { 0x00000200, 0x00000200, "EO" },
+    { 0x00000100, 0x00000100, "EU" },
+    { 0x00000080, 0x00000080, "EI" },
+    { 0x00000040, 0x00000040, "FV" },
+    { 0x00000020, 0x00000020, "FZ" },
+    { 0x00000010, 0x00000010, "FO" },
+    { 0x00000008, 0x00000008, "FU" },
+    { 0x00000004, 0x00000004, "FI" },
+    { 0x00000003, 0x00000000, "RN" },
+    { 0x00000003, 0x00000001, "RZ" },
+    { 0x00000003, 0x00000002, "RP" },
+    { 0x00000003, 0x00000003, "RM" },
+    { 0x00000000, 0x00000000, "" },
+};
+char *crash_screen_strings[] = {
+    "setup",
+    "control",
+    "print",
+    "update",
+    "free"
+};
+s32 D_800937F0 = 0;
+s32 gSomeCrashVideoFlag = 0;
+s32 D_800937f8[3] = {0};
+u8 gCFileLabelIndex = 0;
+u8 gCFileLabelFlag = 0;
+
+void get_err_string(s32 x, s32 y, u32 param3, CrashErrString *param4);
 void crash_copy_control_inputs();
 void crash_print_line(s32 x, s32 y, char *fmt, ...);
 
@@ -132,12 +280,6 @@ void crash_controller_getter() {
 
 #pragma GLOBAL_ASM("asm/nonmatchings/exception/print_stack_trace.s")
 
-extern char D_8009B5D4[]; // "FPU INFO"
-extern char D_8009B5E0[]; // "Fault in thread %d"
-extern char D_8009B5F4[]; // "fpcsr : %08x"
-extern char D_8009B604[]; // "FPU Register dump diabled"
-extern char D_8009B620[]; // "press button for CPU registers"
-
 void other_crash_print(OSThread **threads, s32 count, s32 threadIdx) {
     OSThread *thread;
     __OSThreadContext *context;
@@ -148,20 +290,20 @@ void other_crash_print(OSThread **threads, s32 count, s32 threadIdx) {
     clear_framebuffer_current();
 
     D_800937F0 = 6;
-    crash_print_line(244, 214, D_8009B5D4);
+    crash_print_line(244, 214, "FPU INFO");
 
     D_800937F0 = 0;
-    crash_print_line(16, 24, D_8009B5E0, thread->id);
+    crash_print_line(16, 24, "Fault in thread %d", thread->id);
 
     D_800937F0 = 3;
-    crash_print_line(16, 34, D_8009B5F4, context->fpcsr);
+    crash_print_line(16, 34, "fpcsr : %08x", context->fpcsr);
     get_err_string(16, 40, context->fpcsr, &errStringArray_cause[59]);
 
     D_800937F0 = 2;
-    crash_print_line(16, 50, D_8009B604);
+    crash_print_line(16, 50, "FPU Register dump diabled");
 
     D_800937F0 = 0;
-    crash_print_line(80, 220, D_8009B620);
+    crash_print_line(80, 220, "press button for CPU registers");
 
     while (1) {
         do {
@@ -269,8 +411,6 @@ void func_80062D38(s32 col, s32 row, char *param3) {
 }
 #endif
 
-extern char D_80093200[]; // TODO: is this really the start of the array?
-
 void func_80062E38(s32 x, s32 y, char *str) {
     char c;
 
@@ -321,19 +461,14 @@ void crash_print_line(s32 x, s32 y, char *fmt, ...) {
     func_80062E38(x, y, str);
 }
 
-extern char D_8009B640[]; // "("
-extern char D_8009B644[]; // ","
-extern char D_8009B648[]; // "%s"
-extern char D_8009B64C[]; // ")"
-
-void get_err_string(s32 x, s32 y, u32 param3, ErrString *param4) {
+void get_err_string(s32 x, s32 y, u32 param3, CrashErrString *param4) {
     s32 bvar;
     char str[260]; // length is anywhere between 253-260
     s32 len;
 
     bvar = 1;
 
-    sprintf(str, D_8009B640);
+    sprintf(str, "(");
     len = strlen(str);
 
     while (param4->code1 != 0) {
@@ -341,21 +476,25 @@ void get_err_string(s32 x, s32 y, u32 param3, ErrString *param4) {
             if (bvar) {
                 bvar = 0;
             } else {
-                sprintf(str + len, D_8009B644);
+                sprintf(str + len, ",");
             }
 
             len = strlen(str);
-            sprintf(str + len, D_8009B648, param4->text);
+            sprintf(str + len, "%s", param4->text);
             len = strlen(str);
         }
 
         param4++;
     }
 
-    sprintf(str + len, D_8009B64C);
+    sprintf(str + len, ")");
     len = strlen(str);
     crash_print_line(x, y, str);
 }
+
+static const char str_8009b650[] = "recent ";
+static const char str_8009b658[] = "       ";
+static const char str_8009b660[] = "%s:%d\n";
 
 /**
  * Sets all values of gFramebufferCurrent to 0.
