--- src/segment_12320.c
@@ -1,29 +1,0 @@
-#include "common.h"
-
-#pragma GLOBAL_ASM("asm/nonmatchings/segment_12320/init_audio.s")
-
-#pragma GLOBAL_ASM("asm/nonmatchings/segment_12320/func_80011AFC.s")
-
-#pragma GLOBAL_ASM("asm/nonmatchings/segment_12320/start_alSyn_thread.s")
-
-#pragma GLOBAL_ASM("asm/nonmatchings/segment_12320/stop_alSyn_thread.s")
-
-#pragma GLOBAL_ASM("asm/nonmatchings/segment_12320/func_80011C70.s")
-
-#pragma GLOBAL_ASM("asm/nonmatchings/segment_12320/NOOP_80011ec0.s")
-
-#pragma GLOBAL_ASM("asm/nonmatchings/segment_12320/al_dame_sub.s")
-
-#pragma GLOBAL_ASM("asm/nonmatchings/segment_12320/al_syn_dma.s")
-
-#pragma GLOBAL_ASM("asm/nonmatchings/segment_12320/func_800120C0.s")
-
-#pragma GLOBAL_ASM("asm/nonmatchings/segment_12320/func_800121DC.s")
-
-#pragma GLOBAL_ASM("asm/nonmatchings/segment_12320/func_80012224.s")
-
-#pragma GLOBAL_ASM("asm/nonmatchings/segment_12320/func_80012230.s")
-
-#pragma GLOBAL_ASM("asm/nonmatchings/segment_12320/func_80012348.s")
-
-#pragma GLOBAL_ASM("asm/nonmatchings/segment_12320/func_80012358.s")
