+++ src/segment_535C0.c
@@ -1,5 +1,7 @@
 #include "common.h"
 
+// NOTE: this file has a lot of 64-bit math going on, including 64-bit rodata constants
+
 #pragma GLOBAL_ASM("asm/nonmatchings/segment_535C0/func_800529C0.s")
 
 #pragma GLOBAL_ASM("asm/nonmatchings/segment_535C0/func_80052DC0.s")
