--- src/segment_7B50.c
@@ -1,89 +1,0 @@
-// voxmap.c?
-
-#include "common.h"
-
-#pragma GLOBAL_ASM("asm/nonmatchings/segment_7B50/func_80006F50.s")
-
-#pragma GLOBAL_ASM("asm/nonmatchings/segment_7B50/func_800070FC.s")
-
-#pragma GLOBAL_ASM("asm/nonmatchings/segment_7B50/func_80007178.s")
-
-#pragma GLOBAL_ASM("asm/nonmatchings/segment_7B50/func_80007224.s")
-
-#pragma GLOBAL_ASM("asm/nonmatchings/segment_7B50/func_80007380.s")
-
-#pragma GLOBAL_ASM("asm/nonmatchings/segment_7B50/func_800073CC.s")
-
-#pragma GLOBAL_ASM("asm/nonmatchings/segment_7B50/func_80007468.s")
-
-#pragma GLOBAL_ASM("asm/nonmatchings/segment_7B50/func_800075B0.s")
-
-#pragma GLOBAL_ASM("asm/nonmatchings/segment_7B50/func_80007620.s")
-
-#pragma GLOBAL_ASM("asm/nonmatchings/segment_7B50/func_80007778.s")
-
-#pragma GLOBAL_ASM("asm/nonmatchings/segment_7B50/func_80007A58.s")
-
-#pragma GLOBAL_ASM("asm/nonmatchings/segment_7B50/func_80007CBC.s")
-
-#pragma GLOBAL_ASM("asm/nonmatchings/segment_7B50/func_80007CCC.s")
-
-#pragma GLOBAL_ASM("asm/nonmatchings/segment_7B50/func_80007D08.s")
-
-#pragma GLOBAL_ASM("asm/nonmatchings/segment_7B50/func_80007E2C.s")
-
-#pragma GLOBAL_ASM("asm/nonmatchings/segment_7B50/func_80007EE0.s")
-
-#pragma GLOBAL_ASM("asm/nonmatchings/segment_7B50/func_80008048.s")
-
-#pragma GLOBAL_ASM("asm/nonmatchings/segment_7B50/func_80008524.s")
-
-#pragma GLOBAL_ASM("asm/nonmatchings/segment_7B50/func_8000896C.s")
-
-#pragma GLOBAL_ASM("asm/nonmatchings/segment_7B50/func_80008BE4.s")
-
-#pragma GLOBAL_ASM("asm/nonmatchings/segment_7B50/func_80008D90.s")
-
-#pragma GLOBAL_ASM("asm/nonmatchings/segment_7B50/func_80008DC0.s")
-
-#pragma GLOBAL_ASM("asm/nonmatchings/segment_7B50/func_80008E08.s")
-
-#pragma GLOBAL_ASM("asm/nonmatchings/segment_7B50/func_80008E40.s")
-
-#pragma GLOBAL_ASM("asm/nonmatchings/segment_7B50/func_80008F34.s")
-
-#pragma GLOBAL_ASM("asm/nonmatchings/segment_7B50/func_80009024.s")
-
-#pragma GLOBAL_ASM("asm/nonmatchings/segment_7B50/func_8000921C.s")
-
-#pragma GLOBAL_ASM("asm/nonmatchings/segment_7B50/func_80009528.s")
-
-#pragma GLOBAL_ASM("asm/nonmatchings/segment_7B50/func_800095B0.s")
-
-#pragma GLOBAL_ASM("asm/nonmatchings/segment_7B50/func_800099D0.s")
-
-#pragma GLOBAL_ASM("asm/nonmatchings/segment_7B50/func_80009F5C.s")
-
-#pragma GLOBAL_ASM("asm/nonmatchings/segment_7B50/func_8000A078.s")
-
-#pragma GLOBAL_ASM("asm/nonmatchings/segment_7B50/func_8000A640.s")
-
-#pragma GLOBAL_ASM("asm/nonmatchings/segment_7B50/func_8000A650.s")
-
-#pragma GLOBAL_ASM("asm/nonmatchings/segment_7B50/func_8000A6D8.s")
-
-#pragma GLOBAL_ASM("asm/nonmatchings/segment_7B50/func_8000A76C.s")
-
-#pragma GLOBAL_ASM("asm/nonmatchings/segment_7B50/func_8000A80C.s")
-
-#pragma GLOBAL_ASM("asm/nonmatchings/segment_7B50/func_8000A884.s")
-
-#pragma GLOBAL_ASM("asm/nonmatchings/segment_7B50/func_8000A934.s")
-
-#pragma GLOBAL_ASM("asm/nonmatchings/segment_7B50/func_8000A9AC.s")
-
-#pragma GLOBAL_ASM("asm/nonmatchings/segment_7B50/func_8000AB18.s")
-
-#pragma GLOBAL_ASM("asm/nonmatchings/segment_7B50/func_8000AB84.s")
-
-#pragma GLOBAL_ASM("asm/nonmatchings/segment_7B50/func_8000ACD4.s")
