+++ symbol_addrs.txt
@@ -965,7 +965,8 @@ osscheduler_ = 0x800ae6b8; // size:0x304
 __ull_divremi = 0x8007ee88; // type:func
 rmonMain = 0x800891c0; // type:func
 developer_names = 0x80093820; // size:0x3c
-errStringArray_cause = 0x80093404; // size:0x2c4
+errStringArray_cause = 0x80093404; // size:0x150
+errStringArray_cause2 = 0x80093554; // size:0x174
 errStringArray_fpsr = 0x800936c8; // size:0x114
 crash_screen_strings = 0x800937dc; // size:0x14
 __osBaseTimer = 0x800cee90; // size:0x20
@@ -1540,9 +1541,6 @@ OSMesgQueue_800bcce0 = 0x800bcce0;
 gDisplayHertz = 0x800bce28; // size:0x4
 D_800bccc0 = 0x800bccc0;
 D_800bce60 = 0x800bce60;
-gPalAspectRatio = 0x8009AD64; // size:0x4
-gMPalAspectRatio = 0x8009AD6C; // size:0x4
-gNtscAspectRatio = 0x8009AD70; // size:0x4
 
 weird_resize_copy = 0x8003F1EC; // type:func
 queue_load_texture_proxy = 0x8003CD7C; // type:func
